# internal/web/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: harbrr management API
  version: 0.1.0
  description: >
    harbrr's HTTP surface: the management API (first-run setup, authentication,
    API-key management, indexer and app-sync configuration) plus the entry points of
    the *arr-facing Torznab feed. The Torznab wire contract itself — the XML body
    shapes — is defined by Jackett/Newznab and served by internal/torznab; here we
    document only its HTTP routes, parameters, and content types (see
    docs/architecture.md). Hand-authored, embedded via //go:embed, and guarded by a
    drift test that asserts the spec matches the mounted routes (make test-openapi).

    Authentication: the management API uses a cookie session (after login) or the
    X-API-Key header; the Torznab feed authenticates with an apikey (or passkey)
    query parameter, as *arr clients expect. CSRF: a cookie-authenticated mutating
    request (POST/PUT/PATCH/DELETE) must echo the session's CSRF token in an
    X-CSRF-Token header — the token is minted at login and handed to the client via
    the non-HttpOnly harbrr_csrf cookie and GET /api/auth/me; a missing or wrong
    token is 403. X-API-Key and auth-disabled callers are exempt (see
    docs/security.md).
  license:
    name: GPL-2.0-or-later
    url: https://www.gnu.org/licenses/gpl-2.0.html
...
paths:
...
  /api/indexers/{slug}/search:
    get:
      summary: Search a configured indexer (JSON)
      description: >
        Runs a search against the indexer and returns the same releases the
        Torznab feed serves for the same query (the shared pipeline — parity).
        For a resolver-needing indexer each download link is sealed behind the
        /dl proxy, so a passkey never appears in the response.
      operationId: searchIndexer
      tags: [Indexers]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: q, in: query, required: false, schema: { type: string }, description: free-text query }
        - { name: cat, in: query, required: false, schema: { type: string }, description: comma-separated newznab category ids }
        - { name: imdbid, in: query, required: false, schema: { type: string } }
        - { name: tmdbid, in: query, required: false, schema: { type: string } }
        - { name: tvdbid, in: query, required: false, schema: { type: string } }
        - { name: season, in: query, required: false, schema: { type: string } }
        - { name: ep, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer }, description: "page size, clamped to [1,100]" }
        - { name: offset, in: query, required: false, schema: { type: integer } }
      responses:
        "200":
          description: the search results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResults" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500":
          description: the search failed (a sanitized, secret-free error)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/indexers/{slug}/download/{token}:
    get:
      summary: Download a search result (session-authed)
      description: >
        Resolves and streams a search result's .torrent/.nzb bytes (or 302-redirects a
        magnet), authenticating by session cookie or X-API-Key — the web UI's sibling of
        the feed's apikey-sealed /dl proxy (the UI authenticates by cookie and never
        sends X-API-Key). The token is the opaque, indexer-bound value the JSON search
        response sealed into each resolver-needing release link; the passkey-bearing
        source link is resolved and fetched server-side and never exposed.
      operationId: downloadRelease
      tags: [Indexers]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: token, in: path, required: true, schema: { type: string }, description: opaque download token from the search response link }
      responses:
        "200":
          description: the torrent/nzb file bytes
          content:
            application/x-bittorrent:
              schema: { type: string, format: binary }
            application/x-nzb:
              schema: { type: string, format: binary }
        "302":
          description: a resolved magnet link (Location header); no body
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500":
          description: the grab failed (a sanitized, secret-free error)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: the download proxy is not enabled (no token keyring)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: harbrr_session
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        A bearer credential — send over HTTPS only; it must never traverse a
        cleartext link. Terminate TLS at harbrr or a trusted reverse proxy.
  responses:
    BadRequest:
      description: invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: authentication required or failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: human-readable message }
        code:
          type: string
          description: >
            machine-readable code clients can branch on — one of: bad_request,
            invalid, unauthorized, invalid_credentials, invalid_api_key, forbidden,
            not_found, conflict, already_setup, not_implemented, internal.
      required: [error, code]
...
    SearchResults:
      type: object
      description: >
        A page of search results plus paging metadata (qui-shaped envelope). `total`
        is the full match count after dedupe/filter but before the page slice, so it
        can exceed the number of results returned; `hasMore` is true when results
        beyond this page exist; `limit`/`offset` are the resolved page window. Deep
        server-side multi-page upstream fetching is not yet implemented, so `total`
        reflects the single engine fetch that backs every page of one query.
      properties:
        results:
          type: array
          items: { $ref: "#/components/schemas/Release" }
        total:
          type: integer
          description: full match count before the page slice
        hasMore:
          type: boolean
          description: whether results beyond this page exist (offset + page length < total)
        limit:
          type: integer
          description: resolved page size (clamped to [1,100])
        offset:
          type: integer
          description: resolved page offset
      required: [results, total, hasMore, limit, offset]
    Release:
      type: object
      description: >
        A normalized release — the same shape the Torznab feed serializes. For a
        resolver-needing indexer, `link` is the opaque /dl proxy URL (the passkey is
        sealed inside, never exposed here). Zero values for size/seeders/leechers/
        peers and the volume factors are meaningful and always present.
      properties:
        title: { type: string }
        link: { type: string }
        magnet: { type: string }
        infohash: { type: string }
        size: { type: integer, format: int64 }
        categories: { type: array, items: { type: integer } }
        seeders: { type: integer, format: int64 }
        leechers: { type: integer, format: int64 }
        peers: { type: integer, format: int64 }
        grabs: { type: integer, format: int64 }
        files: { type: integer, format: int64 }
        publishDate: { type: string, format: date-time }
        downloadVolumeFactor: { type: number }
        uploadVolumeFactor: { type: number }
        imdbid: { type: string }
        tmdbid: { type: integer, format: int64 }
        tvdbid: { type: integer, format: int64 }
      required: [title]
...